renamed classes to work with new json structure global
