+++ src/main/java/com/vik/covid19vik/ApiController.java
@@ -1,8 +1,6 @@
 package com.vik.covid19vik;
 
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 // =============== API calls =============== //
@@ -10,32 +8,22 @@ import org.springframework.web.bind.annotation.RestController;
 class ApiController {
     // consider adding to database:
     // --------------- global series data ----------------- //
-    @GetMapping("/API/series/global/confirmed")
-    CountryGlobal[] globalConfirmed() {
+    @GetMapping("API/series/global/confirmed")
+    String global2Confirmed() {
         String globalConfData = JHUPullMethods.getTimeSeriesGlobalConf();
-        CountryGlobal[] confirmedJson = CountryGlobalDataParse.fromJSON("confirmed", globalConfData);
-        return confirmedJson;
+        return CountryGlobalDataParse.parseDataToJSON("confirmed", globalConfData);
     }
 
     @GetMapping("/API/series/global/deaths")
-    CountryGlobal[] globalDeaths() {
+    String global2Deaths() {
         String globalDeathsData = JHUPullMethods.getTimeSeriesGlobalDeaths();
-        CountryGlobal[] deathsJson = CountryGlobalDataParse.fromJSON("deaths", globalDeathsData);
-        return deathsJson;
+        return CountryGlobalDataParse.parseDataToJSON("deaths", globalDeathsData);
     }
 
     @GetMapping("API/series/global/recovered")
-    CountryGlobal[] globalRecovered() {
+    String global2Recovered() {
         String globalRecovData = JHUPullMethods.getTimeSeriesGlobalRecov();
-        CountryGlobal[] recovJson = CountryGlobalDataParse.fromJSON("recovered", globalRecovData);
-        return recovJson;
-    }
-
-    @GetMapping("API/series/global/confirmed2")
-    CountryGlobal2 global2Confirmed() {
-        String globalConfData = JHUPullMethods.getTimeSeriesGlobalConf();
-        CountryGlobal2 confirmedJson = CountryGlobalDataParse2.fromJSON("confirmed", globalConfData);
-        return confirmedJson;
+        return CountryGlobalDataParse.parseDataToJSON("recovered", globalRecovData);
     }
 
 //    // ------------------- US series data ------------------ //
@@ -63,9 +51,6 @@ class ApiController {
     // ------------ UID, IOS, FIPS lookup data ------------- //
     @GetMapping("API/uifcountries")
     String uifLookup() {
-        String json = CountryUIFLookupParse.parseDatatoJSON();
-//        CountryUIFLookup[] uifCountries = CountryUIFLookupParse.fromJSON();
-//        return uifCountries;
-        return json;
+        return CountryUIFLookupParse.parseDatatoJSON();
     }
 }
