+++ src/main/java/com/vik/covid19vik/ApiMethods.java
@@ -1,0 +1,5 @@
+package com.vik.covid19vik;
+
+public class ApiMethods {
+    
+}
